requirements.txt
pillow==10.4.0
pydantic==2.8.2
pyyaml==6.0.2
rich==13.7.1
openai==1.51.0
httpx==0.27.2
